requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic-settings==2.2.1
apscheduler==3.10.4
aiohttp==3.9.1
python-dotenv==1.0.0
pytz==2023.3